Battery Health Charging: Remove check function from installer.
Current code works on ubuntu which uses legacy version of polkit. So 'check' command works successfull without root privileges.
But when a user tested on Archlinux which uses the latest version of polkit (non legacy).
'check' command failed notifying users to update polkit everytime it initialize.
Have to find a different way to check if rules, policies, and ctl is current with the extension version.