Battery health Charging: Correct Battery status icon in system tray
When charging threshold is set, Upower reports Device.state as either charging,
charge-pending or discharging depending on the battery level.

This displays the battery Status icon sometimes to on-battery even though charger is always plugged-in.

There is no way of knowing that charger is powered-off / disconnected, if we miss the notification sound

We will read powersupply AC/ADP online status and change the battery icon accordingly,
So that when charger is plugged (regardless of battery not charging due the charging threshold set) battery icon will display as charging/plugged-in.